fix: pack checkpoint (#59)
* Implementation of weightedPartialFillRate

* fix tests

* fix old tests

* Pass some test

* before refactor

* Fix up

* Issue w lowerCheckpoint

* pre-commit

* fix conflicts

* simplify exitPartiallyFilledBid

* Fix lastFullyFilledCheckpoint hint validation

* pre-commit

* Fix prev and next pointers in CheckpointStorage

* Add fuzz

* Simplify math

* Update tests all passing and pre-commit

* update pre-commit

* update readme

* feedback

* save

* Mass replace uint256 for uint128

* save

* Fix invariant and update pre-commit

* pre-commit

* fix interface

* Remove unused library for 256

+++ src/BidStorage.sol
@@ -22,7 +22,7 @@ abstract contract BidStorage {
     /// @param owner The owner of the bid
     /// @param maxPrice The maximum price for the bid
     /// @return bidId The id of the created bid
-    function _createBid(bool exactIn, uint256 amount, address owner, uint256 maxPrice)
+    function _createBid(bool exactIn, uint128 amount, address owner, uint256 maxPrice)
         internal
         returns (uint256 bidId)
     {
