+++ src/interfaces/external/IDistributionStrategy.sol
@@ -16,7 +16,7 @@ interface IDistributionStrategy {
     /// @param salt The salt to use for the deterministic deployment.
     /// @return distributionContract The contract that will handle or manage the distribution.
     ///         (Could be `address(this)` if the strategy is handled in-place, or a newly deployed instance).
-    function initializeDistribution(address token, uint256 amount, bytes calldata configData, bytes32 salt)
+    function initializeDistribution(address token, uint128 amount, bytes calldata configData, bytes32 salt)
         external
         returns (IDistributionContract distributionContract);
 }
