+++ src/libraries/AuctionStepLib.sol
@@ -30,7 +30,7 @@ library AuctionStepLib {
 
     /// @notice Apply mps to a value
     /// @dev Requires that value is > MPS to avoid loss of precision
-    function applyMps(uint256 value, uint24 mps) internal pure returns (uint256) {
+    function applyMps(uint128 value, uint24 mps) internal pure returns (uint128) {
         return mps * value / MPS;
     }
 }
