+++ test/Auction.t.sol
@@ -24,7 +24,7 @@ import {FixedPointMathLib} from 'solady/utils/FixedPointMathLib.sol';
 import {SafeTransferLib} from 'solady/utils/SafeTransferLib.sol';
 
 contract AuctionTest is AuctionBaseTest {
-    using FixedPointMathLib for uint256;
+    using FixedPointMathLib for uint128;
     using AuctionParamsBuilder for AuctionParameters;
     using AuctionStepsBuilder for bytes;
 
@@ -33,8 +33,8 @@ contract AuctionTest is AuctionBaseTest {
     }
 
     /// Return the inputAmount required to purchase at least the given number of tokens at the given maxPrice
-    function inputAmountForTokens(uint256 tokens, uint256 maxPrice) internal pure returns (uint256) {
-        return tokens.fullMulDivUp(maxPrice, FixedPoint96.Q96);
+    function inputAmountForTokens(uint128 tokens, uint256 maxPrice) internal pure returns (uint128) {
+        return uint128(tokens.fullMulDivUp(maxPrice, FixedPoint96.Q96));
     }
 
     /// forge-config: default.isolate = true
@@ -94,7 +94,7 @@ contract AuctionTest is AuctionBaseTest {
         vm.snapshotGasLastCall('submitBid_recordStep_updateCheckpoint_initializeTick');
 
         vm.roll(block.number + 1);
-        uint256 expectedTotalCleared = 10e18; // 100e3 mps * total supply (1000e18)
+        uint128 expectedTotalCleared = 10e18; // 100e3 mps * total supply (1000e18)
         uint24 expectedCumulativeMps = 100e3; // 100e3 mps * 1 block
         vm.expectEmit(true, true, true, true);
         emit IAuction.CheckpointUpdated(
@@ -114,7 +114,7 @@ contract AuctionTest is AuctionBaseTest {
         );
 
         vm.roll(block.number + 1);
-        uint256 expectedTotalCleared = 10e18; // 100e3 mps * total supply (1000e18)
+        uint128 expectedTotalCleared = 10e18; // 100e3 mps * total supply (1000e18)
         uint24 expectedCumulativeMps = 100e3; // 100e3 mps * 1 block
         vm.expectEmit(true, true, true, true);
         emit IAuction.CheckpointUpdated(
@@ -141,7 +141,7 @@ contract AuctionTest is AuctionBaseTest {
 
         vm.roll(block.number + 1);
         uint24 expectedCumulativeMps = 100e3; // 100e3 mps * 1 block
-        uint256 expectedTotalCleared = 10e18; // 100e3 mps * total supply (1000e18)
+        uint128 expectedTotalCleared = 10e18; // 100e3 mps * total supply (1000e18)
         vm.expectEmit(true, true, true, true);
         emit IAuction.CheckpointUpdated(
             block.number, tickNumberToPriceX96(2), expectedTotalCleared, expectedCumulativeMps
@@ -150,7 +150,7 @@ contract AuctionTest is AuctionBaseTest {
     }
 
     function test_submitBid_multipleTicks_succeeds() public {
-        uint256 expectedTotalCleared = 100e3 * TOTAL_SUPPLY / AuctionStepLib.MPS;
+        uint128 expectedTotalCleared = 100e3 * TOTAL_SUPPLY / AuctionStepLib.MPS;
         uint24 expectedCumulativeMps = 100e3; // 100e3 mps * 1 block
 
         vm.expectEmit(true, true, true, true);
@@ -192,7 +192,7 @@ contract AuctionTest is AuctionBaseTest {
     }
 
     function test_submitBid_exactIn_overTotalSupply_isPartiallyFilled() public {
-        uint256 inputAmount = inputAmountForTokens(2000e18, tickNumberToPriceX96(2));
+        uint128 inputAmount = inputAmountForTokens(2000e18, tickNumberToPriceX96(2));
         uint256 bidId = auction.submitBid{value: inputAmount}(
             tickNumberToPriceX96(2), true, inputAmount, alice, tickNumberToPriceX96(1), bytes('')
         );
@@ -244,7 +244,7 @@ contract AuctionTest is AuctionBaseTest {
         token.mint(address(auction), TOTAL_SUPPLY);
 
         // Bid over the total supply
-        uint256 inputAmount = inputAmountForTokens(2000e18, tickNumberToPriceX96(2));
+        uint128 inputAmount = inputAmountForTokens(2000e18, tickNumberToPriceX96(2));
         vm.expectEmit(true, true, true, true);
         emit IAuction.CheckpointUpdated(block.number, 0, 0, 0);
         vm.expectEmit(true, true, true, true);
@@ -263,7 +263,7 @@ contract AuctionTest is AuctionBaseTest {
         // Advance to the end of the first step
         vm.roll(auction.startBlock() + 101);
 
-        uint256 expectedTotalCleared = 100e3 * TOTAL_SUPPLY / AuctionStepLib.MPS;
+        uint128 expectedTotalCleared = 100e3 * TOTAL_SUPPLY / AuctionStepLib.MPS;
         // Now the auction should start clearing
         vm.expectEmit(true, true, true, true);
         emit IAuction.CheckpointUpdated(block.number, tickNumberToPriceX96(2), expectedTotalCleared, 100e3);
@@ -288,7 +288,7 @@ contract AuctionTest is AuctionBaseTest {
         auction = new Auction(address(token), TOTAL_SUPPLY, params);
         token.mint(address(auction), TOTAL_SUPPLY);
 
-        uint256 inputAmount = inputAmountForTokens(1000e18, tickNumberToPriceX96(1));
+        uint128 inputAmount = inputAmountForTokens(1000e18, tickNumberToPriceX96(1));
         vm.expectEmit(true, true, true, true);
         emit IAuction.CheckpointUpdated(block.number, 0, 0, 0);
         vm.expectEmit(true, true, true, true);
@@ -308,7 +308,7 @@ contract AuctionTest is AuctionBaseTest {
         // Advance to the end of the first step
         vm.roll(auction.startBlock() + 101);
 
-        uint256 expectedTotalCleared = 100e3 * TOTAL_SUPPLY / AuctionStepLib.MPS;
+        uint128 expectedTotalCleared = 100e3 * TOTAL_SUPPLY / AuctionStepLib.MPS;
         // Now the auction should start clearing
         vm.expectEmit(true, true, true, true);
         emit IAuction.CheckpointUpdated(block.number, tickNumberToPriceX96(1), expectedTotalCleared, 100e3);
@@ -401,7 +401,7 @@ contract AuctionTest is AuctionBaseTest {
     /// forge-config: default.isolate = true
     /// forge-config: ci.isolate = true
     function test_exitBid_succeeds() public {
-        uint256 smallAmount = 500e18;
+        uint128 smallAmount = 500e18;
         vm.expectEmit(true, true, true, true);
         emit IAuction.BidSubmitted(
             0, alice, tickNumberToPriceX96(2), true, inputAmountForTokens(smallAmount, tickNumberToPriceX96(2))
@@ -416,7 +416,7 @@ contract AuctionTest is AuctionBaseTest {
         );
 
         // Bid enough tokens to move the clearing price to 3
-        uint256 largeAmount = 1000e18;
+        uint128 largeAmount = 1000e18;
         vm.expectEmit(true, true, true, true);
         emit IAuction.BidSubmitted(
             1, alice, tickNumberToPriceX96(3), true, inputAmountForTokens(largeAmount, tickNumberToPriceX96(3))
@@ -429,7 +429,7 @@ contract AuctionTest is AuctionBaseTest {
             tickNumberToPriceX96(2),
             bytes('')
         );
-        uint256 expectedTotalCleared = TOTAL_SUPPLY * 100e3 / AuctionStepLib.MPS;
+        uint128 expectedTotalCleared = TOTAL_SUPPLY * 100e3 / AuctionStepLib.MPS;
 
         vm.roll(block.number + 1);
         vm.expectEmit(true, true, true, true);
@@ -453,7 +453,7 @@ contract AuctionTest is AuctionBaseTest {
         auction.exitBid(bidId2);
         vm.stopPrank();
 
-        uint256 expectedCurrentRaised = inputAmountForTokens(largeAmount, tickNumberToPriceX96(3));
+        uint128 expectedCurrentRaised = inputAmountForTokens(largeAmount, tickNumberToPriceX96(3));
         vm.startPrank(auction.fundsRecipient());
         vm.expectEmit(true, true, true, true);
         emit ITokenCurrencyStorage.CurrencySwept(auction.fundsRecipient(), expectedCurrentRaised);
@@ -467,7 +467,7 @@ contract AuctionTest is AuctionBaseTest {
     }
 
     function test_exitBid_exactOut_succeeds() public {
-        uint256 amount = 500e18;
+        uint128 amount = 500e18;
         uint256 maxPrice = tickNumberToPriceX96(2);
         uint256 bidId = auction.submitBid{value: inputAmountForTokens(500e18, tickNumberToPriceX96(2))}(
             maxPrice, false, 500e18, alice, tickNumberToPriceX96(1), bytes('')
@@ -744,7 +744,7 @@ contract AuctionTest is AuctionBaseTest {
 
         // Roll to end of auction
         vm.roll(auction.endBlock());
-        uint256 expectedCurrencyRaised = inputAmountForTokens(750e18, tickNumberToPriceX96(11))
+        uint128 expectedCurrencyRaised = inputAmountForTokens(750e18, tickNumberToPriceX96(11))
             + inputAmountForTokens(100e18, tickNumberToPriceX96(11))
             + inputAmountForTokens(150e18, tickNumberToPriceX96(11));
 
@@ -933,7 +933,7 @@ contract AuctionTest is AuctionBaseTest {
         // The formula is: clearingPrice = currencyDemand * Q96 / (blockTokenSupply - tokenDemand)
         // We want: minimumClearingPrice < calculated_price < floorPrice
         // With currencyDemand = 120e18 and tokenDemand = 100e18, we need to find the right blockTokenSupply
-        uint256 blockTokenSupply = 1e22; // Even larger supply to get a smaller calculated price
+        uint128 blockTokenSupply = 1e22; // Even larger supply to get a smaller calculated price
 
         uint256 result = mockAuction.calculateNewClearingPrice(
             minimumClearingPrice, // minimumClearingPrice in X96 (below floor price)
@@ -1274,7 +1274,7 @@ contract AuctionTest is AuctionBaseTest {
         token.mint(address(auctionWithThreshold), TOTAL_SUPPLY);
 
         // Submit a small bid (only 10% of supply, below 50% threshold)
-        uint256 smallAmount = TOTAL_SUPPLY / 10;
+        uint128 smallAmount = TOTAL_SUPPLY / 10;
         auctionWithThreshold.submitBid{value: inputAmountForTokens(smallAmount, tickNumberToPriceX96(2))}(
             tickNumberToPriceX96(2),
             true,
@@ -1299,8 +1299,8 @@ contract AuctionTest is AuctionBaseTest {
         token.mint(address(auctionWithThreshold), TOTAL_SUPPLY);
 
         // Submit a bid for 50% of supply (above 30% threshold)
-        uint256 halfSupply = TOTAL_SUPPLY / 2;
-        uint256 inputAmount = inputAmountForTokens(halfSupply, tickNumberToPriceX96(2));
+        uint128 halfSupply = TOTAL_SUPPLY / 2;
+        uint128 inputAmount = inputAmountForTokens(halfSupply, tickNumberToPriceX96(2));
         auctionWithThreshold.submitBid{value: inputAmount}(
             tickNumberToPriceX96(2), true, inputAmount, alice, tickNumberToPriceX96(1), bytes('')
         );
@@ -1330,8 +1330,8 @@ contract AuctionTest is AuctionBaseTest {
         token.mint(address(auctionWithCallback), TOTAL_SUPPLY);
 
         // Submit a bid for 50% of supply (above 30% threshold)
-        uint256 halfSupply = TOTAL_SUPPLY / 2;
-        uint256 inputAmount = inputAmountForTokens(halfSupply, tickNumberToPriceX96(2));
+        uint128 halfSupply = TOTAL_SUPPLY / 2;
+        uint128 inputAmount = inputAmountForTokens(halfSupply, tickNumberToPriceX96(2));
         auctionWithCallback.submitBid{value: inputAmount}(
             tickNumberToPriceX96(2), true, inputAmount, alice, tickNumberToPriceX96(1), bytes('')
         );
@@ -1363,8 +1363,8 @@ contract AuctionTest is AuctionBaseTest {
         token.mint(address(auctionWithCallback), TOTAL_SUPPLY);
 
         // Submit a bid for 50% of supply (above 30% threshold)
-        uint256 halfSupply = TOTAL_SUPPLY / 2;
-        uint256 inputAmount = inputAmountForTokens(halfSupply, tickNumberToPriceX96(2));
+        uint128 halfSupply = TOTAL_SUPPLY / 2;
+        uint128 inputAmount = inputAmountForTokens(halfSupply, tickNumberToPriceX96(2));
         auctionWithCallback.submitBid{value: inputAmount}(
             tickNumberToPriceX96(2), true, inputAmount, alice, tickNumberToPriceX96(1), bytes('')
         );
@@ -1388,8 +1388,8 @@ contract AuctionTest is AuctionBaseTest {
         token.mint(address(auctionWithCallback), TOTAL_SUPPLY);
 
         // Submit a bid for 50% of supply (above 30% threshold)
-        uint256 halfSupply = TOTAL_SUPPLY / 2;
-        uint256 inputAmount = inputAmountForTokens(halfSupply, tickNumberToPriceX96(2));
+        uint128 halfSupply = TOTAL_SUPPLY / 2;
+        uint128 inputAmount = inputAmountForTokens(halfSupply, tickNumberToPriceX96(2));
         auctionWithCallback.submitBid{value: inputAmount}(
             tickNumberToPriceX96(2), true, inputAmount, alice, tickNumberToPriceX96(1), bytes('')
         );
@@ -1413,8 +1413,8 @@ contract AuctionTest is AuctionBaseTest {
         token.mint(address(auctionWithCallback), TOTAL_SUPPLY);
 
         // Submit a bid for 50% of supply (above 30% threshold)
-        uint256 halfSupply = TOTAL_SUPPLY / 2;
-        uint256 inputAmount = inputAmountForTokens(halfSupply, tickNumberToPriceX96(2));
+        uint128 halfSupply = TOTAL_SUPPLY / 2;
+        uint128 inputAmount = inputAmountForTokens(halfSupply, tickNumberToPriceX96(2));
         auctionWithCallback.submitBid{value: inputAmount}(
             tickNumberToPriceX96(2), true, inputAmount, alice, tickNumberToPriceX96(1), bytes('')
         );
@@ -1442,8 +1442,8 @@ contract AuctionTest is AuctionBaseTest {
         token.mint(address(auctionWithCallback), TOTAL_SUPPLY);
 
         // Submit a bid for 50% of supply (above 30% threshold)
-        uint256 halfSupply = TOTAL_SUPPLY / 2;
-        uint256 inputAmount = inputAmountForTokens(halfSupply, tickNumberToPriceX96(2));
+        uint128 halfSupply = TOTAL_SUPPLY / 2;
+        uint128 inputAmount = inputAmountForTokens(halfSupply, tickNumberToPriceX96(2));
         auctionWithCallback.submitBid{value: inputAmount}(
             tickNumberToPriceX96(2), true, inputAmount, alice, tickNumberToPriceX96(1), bytes('')
         );
@@ -1475,8 +1475,8 @@ contract AuctionTest is AuctionBaseTest {
         token.mint(address(auctionWithCallback), TOTAL_SUPPLY);
 
         // Submit a bid for 50% of supply (above 30% threshold)
-        uint256 halfSupply = TOTAL_SUPPLY / 2;
-        uint256 inputAmount = inputAmountForTokens(halfSupply, tickNumberToPriceX96(2));
+        uint128 halfSupply = TOTAL_SUPPLY / 2;
+        uint128 inputAmount = inputAmountForTokens(halfSupply, tickNumberToPriceX96(2));
         auctionWithCallback.submitBid{value: inputAmount}(
             tickNumberToPriceX96(2), true, inputAmount, alice, tickNumberToPriceX96(1), bytes('')
         );
@@ -1517,8 +1517,8 @@ contract AuctionTest is AuctionBaseTest {
         token.mint(address(secondAuction), TOTAL_SUPPLY);
 
         // Submit bids to both auctions
-        uint256 halfSupply = TOTAL_SUPPLY / 2;
-        uint256 inputAmount = inputAmountForTokens(halfSupply, tickNumberToPriceX96(2));
+        uint128 halfSupply = TOTAL_SUPPLY / 2;
+        uint128 inputAmount = inputAmountForTokens(halfSupply, tickNumberToPriceX96(2));
 
         firstAuction.submitBid{value: inputAmount}(
             tickNumberToPriceX96(2), true, inputAmount, alice, tickNumberToPriceX96(1), bytes('')
@@ -1568,8 +1568,8 @@ contract AuctionTest is AuctionBaseTest {
         token.mint(address(auctionWithThreshold), TOTAL_SUPPLY);
 
         // Submit a bid for 60% of supply (above 30% threshold, so graduated)
-        uint256 soldAmount = (TOTAL_SUPPLY * 60) / 100;
-        uint256 inputAmount = inputAmountForTokens(soldAmount, tickNumberToPriceX96(1));
+        uint128 soldAmount = (TOTAL_SUPPLY * 60) / 100;
+        uint128 inputAmount = inputAmountForTokens(soldAmount, tickNumberToPriceX96(1));
         auctionWithThreshold.submitBid{value: inputAmount}(
             tickNumberToPriceX96(2), true, inputAmount, alice, tickNumberToPriceX96(1), bytes('')
         );
@@ -1579,7 +1579,7 @@ contract AuctionTest is AuctionBaseTest {
         auctionWithThreshold.checkpoint();
 
         // Should sweep only unsold tokens (40% of supply)
-        uint256 expectedUnsoldTokens = TOTAL_SUPPLY - soldAmount;
+        uint128 expectedUnsoldTokens = TOTAL_SUPPLY - soldAmount;
 
         vm.expectEmit(true, true, true, true);
         emit ITokenCurrencyStorage.TokensSwept(tokensRecipient, expectedUnsoldTokens);
@@ -1596,8 +1596,8 @@ contract AuctionTest is AuctionBaseTest {
         token.mint(address(auctionWithThreshold), TOTAL_SUPPLY);
 
         // Submit a small bid for 10% of supply (below 50% threshold, so not graduated)
-        uint256 smallAmount = TOTAL_SUPPLY / 10;
-        uint256 inputAmount = inputAmountForTokens(smallAmount, tickNumberToPriceX96(1));
+        uint128 smallAmount = TOTAL_SUPPLY / 10;
+        uint128 inputAmount = inputAmountForTokens(smallAmount, tickNumberToPriceX96(1));
         auctionWithThreshold.submitBid{value: inputAmount}(
             tickNumberToPriceX96(2), true, inputAmount, alice, tickNumberToPriceX96(1), bytes('')
         );
@@ -1623,8 +1623,8 @@ contract AuctionTest is AuctionBaseTest {
         token.mint(address(auctionWithThreshold), TOTAL_SUPPLY);
 
         // Submit a bid for 70% of supply (above threshold)
-        uint256 soldAmount = (TOTAL_SUPPLY * 70) / 100;
-        uint256 inputAmount = inputAmountForTokens(soldAmount, tickNumberToPriceX96(1));
+        uint128 soldAmount = (TOTAL_SUPPLY * 70) / 100;
+        uint128 inputAmount = inputAmountForTokens(soldAmount, tickNumberToPriceX96(1));
         auctionWithThreshold.submitBid{value: inputAmount}(
             tickNumberToPriceX96(2), true, inputAmount, alice, tickNumberToPriceX96(1), bytes('')
         );
@@ -1634,14 +1634,14 @@ contract AuctionTest is AuctionBaseTest {
         auctionWithThreshold.checkpoint();
 
         // Sweep currency first (should succeed as graduated)
-        uint256 expectedCurrencyRaised = inputAmountForTokens(soldAmount, tickNumberToPriceX96(1));
+        uint128 expectedCurrencyRaised = inputAmountForTokens(soldAmount, tickNumberToPriceX96(1));
         vm.prank(fundsRecipient);
         vm.expectEmit(true, true, true, true);
         emit ITokenCurrencyStorage.CurrencySwept(fundsRecipient, expectedCurrencyRaised);
         auctionWithThreshold.sweepCurrency();
 
         // Then sweep unsold tokens
-        uint256 expectedUnsoldTokens = TOTAL_SUPPLY - soldAmount;
+        uint128 expectedUnsoldTokens = TOTAL_SUPPLY - soldAmount;
         vm.expectEmit(true, true, true, true);
         emit ITokenCurrencyStorage.TokensSwept(tokensRecipient, expectedUnsoldTokens);
         auctionWithThreshold.sweepUnsoldTokens();
@@ -1659,8 +1659,8 @@ contract AuctionTest is AuctionBaseTest {
         token.mint(address(auctionWithThreshold), TOTAL_SUPPLY);
 
         // Submit a bid for only 20% of supply (below 80% threshold)
-        uint256 smallAmount = TOTAL_SUPPLY / 5;
-        uint256 inputAmount = inputAmountForTokens(smallAmount, tickNumberToPriceX96(1));
+        uint128 smallAmount = TOTAL_SUPPLY / 5;
+        uint128 inputAmount = inputAmountForTokens(smallAmount, tickNumberToPriceX96(1));
         auctionWithThreshold.submitBid{value: inputAmount}(
             tickNumberToPriceX96(2), true, inputAmount, alice, tickNumberToPriceX96(1), bytes('')
         );
