+++ test/AuctionFactory.t.sol
@@ -21,7 +21,7 @@ contract AuctionFactoryTest is TokenHandler, Test {
     uint256 public constant AUCTION_DURATION = 100;
     uint256 public constant TICK_SPACING = 1e6;
     uint128 public constant FLOOR_PRICE = 1e6;
-    uint256 public constant TOTAL_SUPPLY = 1000e18;
+    uint128 public constant TOTAL_SUPPLY = 1000e18;
 
     address public alice;
     address public tokensRecipient;
